Allows creating dataset without the need to specify resource
Now in the second stage of creating dataset, it's needed to specify at least one resource. Otherwise it's not possible to continue to the next stage of creating dataset.

This enhancement adds new property "ckan.dataset_form.require_resources". If its not specified or has value "True", CKAN behaves normally (requires resource). If it's value is "False", you can continue to the next stage.